🔥 refactor(chat-ia): mover AgentRuntimeErrorType de model-runtime → types
Romper la dep cruzada types ↔ model-runtime que forzaba al bundler
cargar todo el package model-runtime (284 archivos, SDKs OpenAI/
Anthropic/Bedrock/...) cuando un archivo solo necesita el const
AgentRuntimeErrorType (47 entries enum-like).

Cambios:
- types/src/agentRuntimeError.ts (NUEVO): fuente de verdad de
  AgentRuntimeErrorType, AGENT_RUNTIME_ERROR_SET,
  ILobeAgentRuntimeErrorType.
- types/src/index.ts: re-export del nuevo módulo.
- types/src/fetch.ts y types/src/message/common/base.ts: import
  local en vez de @lobechat/model-runtime.
- model-runtime/src/types/error.ts: re-exporta desde types (backward
  compat — consumidores existentes siguen funcionando).

Próximo paso: migrar consumidores chat-ia/src que solo usan
AgentRuntimeErrorType para que importen desde @lobechat/types
(bundler hace tree-shake real → no carga model-runtime).
