+++ apps/chat-ia/packages/types/src/message/common/base.ts
@@ -1,4 +1,4 @@
-import type { ILobeAgentRuntimeErrorType } from '@lobechat/model-runtime';
+import type { ILobeAgentRuntimeErrorType } from '../../agentRuntimeError';
 import type { IPluginErrorType } from '@lobehub/chat-plugin-sdk';
 
 import { ErrorType } from '../../fetch';
